% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{as_character_inner}
\alias{as_character_inner}
\title{Private function for converting an emphatic data.frame or matrix to a string}
\usage{
as_character_inner(
  m,
  text,
  fill,
  text_mode = "contrast",
  text_contrast = 1,
  full_colour = FALSE,
  dark_mode = TRUE,
  underline_header = TRUE
)
}
\arguments{
\item{m}{character matrix}

\item{text, fill}{matrices of colours to apply to each given cell.
Dimensions must match that of \code{m}}

\item{text_mode}{How to handle textif no text colour has been
explicitly specified by the user.
\describe{
\item{contrast}{(default) automatically select a contrasting colour for enhanced readability.}
\item{asis}{render text in the default text
     colour for the output device, unless the user has already specified
     a text colour at this location}
\item{remove}{remove all text without a user-defined colour}
}}

\item{text_contrast}{When \code{text_mode='contrast'} this numeric value in
range [0, 1] adjusts the visibility. Default: 1 (high contrast)}

\item{full_colour}{Use 24bit ANSI escape codes?  default: FALSE - use 8bit colour.
Note: RStudio only supports 8 bit ANSI output (24bit ANSI is
rendered invisibly in Rstudio).  For 24bit colour output, try R in the terminal
e.g. 'iTerm' on OSX.}

\item{dark_mode}{Output terminal is in 'dark mode'? default: TRUE means that
the terminal display is light coloured text on a dark background.
If your terminal displays dark text on a light background, set
\code{dark_mode = FALSE}}

\item{underline_header}{Draw an underline separating the column header from
the data? Default: TRUE}
}
\description{
Private function for converting an emphatic data.frame or matrix to a string
}
