% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{as_html}
\alias{as_html}
\title{Convert ANSI to HTML}
\usage{
as_html(x, style = NULL, ...)
}
\arguments{
\item{x}{emphatic object}

\item{style}{html tag styling to apply to the \code{<pre>} wrapper for the
returned HTML}

\item{...}{other arguments passed to \code{as.character.emphatic}}
}
\description{
Convert ANSI to HTML
}
