% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-latex.R
\name{as_latex}
\alias{as_latex}
\title{Render an emphatic object to Latex}
\usage{
as_latex(x, ...)
}
\arguments{
\item{x}{emphatic object}

\item{...}{other arguments passed to \code{as.character.emphatic}}
}
\description{
Render an emphatic object to Latex
}
