% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-svg.R
\name{as_svg_anim}
\alias{as_svg_anim}
\title{Wrap multiple emphatic object into an SVG animation}
\usage{
as_svg_anim(
  x,
  width = 1200,
  height = 900,
  duration = 1,
  playback = c("infinite", "click"),
  svg_id = NULL,
  browsable = FALSE
)
}
\arguments{
\item{x}{list of emphatic objects}

\item{width, height}{viewBox dimensions for SVG}

\item{duration}{frame duration in seconds. May be a single value used for
all frames, or a vector of values (one duration value for each frame).
Can be fractions of a second.}

\item{playback}{'click', 'infinite'}

\item{svg_id}{ID to use for the SVG tag.  Default: NULL means to create a random
ID}

\item{browsable}{Should the SVG be rendered to the RStudio Viewer pane when
when printed (instead of console output)? Default: FALSE}
}
\value{
character string containing an SVG snippet.
}
\description{
Idea borrowed from pointblank
}
