% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{as_typst}
\alias{as_typst}
\title{Render an emphatic object to typst}
\usage{
as_typst(x, ..., font = "Courier New", font_size = 10)
}
\arguments{
\item{x}{emphatic object}

\item{...}{other arguments passed to \code{as.character.emphatic}}

\item{font}{name of font. Default: 'Courier New'}

\item{font_size}{font size in points. default: 10}
}
\description{
Render an emphatic object to typst
}
