% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calc_contrasting_text}
\alias{calc_contrasting_text}
\title{Calculate a contrasting text colours for vector of fill colours}
\usage{
calc_contrasting_text(fill, text_contrast, dark_mode)
}
\arguments{
\item{fill}{vector of background colours}

\item{text_contrast}{When \code{text_mode='contrast'} this numeric value in
range [0, 1] adjusts the visibility. Default: 1 (high contrast)}

\item{dark_mode}{Output terminal is in 'dark mode'? default: TRUE means that
the terminal display is light coloured text on a dark background.
If your terminal displays dark text on a light background, set
\code{dark_mode = FALSE}}
}
\value{
contrasting text colour for readable text
}
\description{
The function calculates the luma of each fill colour and then picks either
the text colour to be either white or black in order to contrast with it.
}
\details{
There is a further call to 'interp_colour' in order to control the contrast.
}
