% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sets.R
\docType{data}
\name{challenger}
\alias{challenger}
\title{Challenger o-ring dataset}
\format{
A data.frame
\describe{
  \item{flight}{Flight number}
  \item{temp}{Launch temperature (Fahrenheit)}
  \item{erosion}{Number of o-ring erosion incidents}
  \item{blowby}{Number of o-ring blow-by incidents}
  \item{damage}{Damage severity index}
  \item{date}{Date of launch}
}

Other references:

\itemize{
\item{\url{http://williamwolff.org/wp-content/uploads/2013/01/tufte-challenger-1997.pdf}}
\item{\url{https://archive.ics.uci.edu/ml/datasets/Challenger+USA+Space+Shuttle+O-Ring}}
}
}
\source{
\url{http://williamwolff.org/wp-content/uploads/2013/01/tufte-challenger-1997.pdf}
}
\usage{
challenger
}
\description{
A dataset containing information about the o-ring status of the flights
leading up to the Space Shuttle Challenger distaster.
}
\details{
Sourced from a table in Tufte's "Visual and Statistical Thinking"
}
\keyword{datasets}
