% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coerce_to_string}
\alias{coerce_to_string}
\title{Convert an arbitrary R object to a string}
\usage{
coerce_to_string(x, coerce)
}
\arguments{
\item{x}{object}

\item{coerce}{How should non-character arguments be coerced to character strings?
\describe{
  \item{default}{ - the given object \code{x} must already be a character string}
  \item{character}{ - performs the matching after first calling
          \code{as.character(x)}}
  \item{print}{ - performs the matching against the default
           \code{print(x)} output}
  \item{deparse}{ - performs the matching after first calling
          \code{deparse1(x)}}
  \item{str}{ - performs the matching on the output of calling
          \code{str(x)}}
}}
}
\description{
Convert an arbitrary R object to a string
}
