% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansi-escape-codes-24.R
\name{col2bg24}
\alias{col2bg24}
\alias{col2fg24}
\title{Convert an R colour to an ansi string for 24bit colour supported by some terminals}
\usage{
col2bg24(rcolour)

col2fg24(rcolour)
}
\arguments{
\item{rcolour}{any R colour e.g. 'red', '#445566'.  If rcolour is NA or empty
string then return an empty string}
}
\value{
ANSI escape string for the given colour as a foreground or background
        colour
}
\description{
Ref: \url{https://en.wikipedia.org/wiki/ANSI_escape_code#8-bit}
}
