% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-ansi.R
\name{col2fill_ansi}
\alias{col2fill_ansi}
\alias{col2text_ansi}
\alias{col2code}
\title{Convert an R colour to a 216-colour ANSI string. Suitable for most terminals}
\usage{
col2fill_ansi(rcolour)

col2text_ansi(rcolour)

col2code(rcolour)
}
\arguments{
\item{rcolour}{any R colour e.g. 'red', '#445566'}
}
\value{
ANSI escape string for the given colour as a foreground or background
        colour
}
\description{
Ref: \url{https://en.wikipedia.org/wiki/ANSI_escape_code#8-bit}
}
\details{
If all the RGB colour components are equal then the colour is matched to one
of 24 grey levels, other wise it is converted to one of 216 standard colours.
}
