% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-html.R
\name{col2fill_html}
\alias{col2fill_html}
\title{Convert a vector of colours to opening html spans for background colour}
\usage{
col2fill_html(colours)
}
\arguments{
\item{colours}{Chcaracter vector of R colours}
}
\value{
Character vector of HTML opening span tags
}
\description{
Convert a vector of colours to opening html spans for background colour
}
