% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-html.R
\name{col2hex}
\alias{col2hex}
\title{Convert R colours to hex colours}
\usage{
col2hex(colours)
}
\arguments{
\item{colours}{Character vector of R colours}
}
\value{
Character vector of 6-char hex colours
}
\description{
Convert R colours to hex colours
}
