% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{hl_inner}
\alias{hl_inner}
\title{Private inner function for highlighting}
\usage{
hl_inner(.data, colour, row_ids, column, dest_col_ids, elem)
}
\arguments{
\item{.data}{\code{emphatic} data.frame}

\item{colour}{colour to use for highlighting.  This may be an R colour,
a vector of R colours, or
a \code{ggplot2} style "Scale" object e.g. \code{scale_colour_continuous()}.}

\item{row_ids}{numeric indices of which rows and columns to highlight}

\item{column}{Single numeric index of the column to colour}

\item{dest_col_ids}{column ids to apply highlighting to}

\item{elem}{Apply the highlighting to the 'fill' (the background) or the 'text'.
Default: 'fill'}
}
\description{
This function assumes a lot of pre-processing has gone on and that the 'rows'
and 'column' are now numeric indices (not expressions, or names or anything else)
}
