% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{hl_loc}
\alias{hl_loc}
\title{Highlight elements by location within an \code{emphatic} data.frame or matrix.}
\usage{
hl_loc(
  .data,
  colour,
  row_ids,
  col_ids,
  elem = "fill",
  major = "row",
  expand_grid = TRUE
)
}
\arguments{
\item{.data}{\code{emphatic} data.frame or matrix.}

\item{colour}{colour to use for highlighting.  This may be an R colour,
a vector of R colours, or
a \code{ggplot2} style "Scale" object e.g. \code{scale_colour_continuous()}.}

\item{row_ids, col_ids}{numeric indices of which rows and columns to highlight}

\item{elem}{Apply the highlighting to the 'fill' (the background) or the 'text'.
Default: 'fill'}

\item{major}{which is the direction in which to first replicate colours?
i.e. should colour be replicated to match the number of rows first,
or the number of columns? Default: 'row'. Possible values: 'row', 'column'}

\item{expand_grid}{How should the vectors of row and column indices be combined
to create location coorinates? If \code{expand_grid = TRUE} (the default)
then a full outer join is done
on rows and cols using \code{expand.grid()} - this means that all possible
combinations of the specified rows and columns will be highlighted.
Otherwise the locations are
created using a simpler call to \code{cbind()}.}
}
\description{
Highlight elements by location within an \code{emphatic} data.frame or matrix.
}
