% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{hl_mat}
\alias{hl_mat}
\alias{hl_vec}
\title{Highlight elements of a matrix or atomic vector}
\usage{
hl_mat(.data, colour, selection = NULL, elem = "fill", show_legend = FALSE)

hl_vec(.data, colour, selection = NULL, elem = "fill", show_legend = FALSE)
}
\arguments{
\item{.data}{\code{emphatic} matrix or atomic vector.}

\item{colour}{colour to use for highlighting.  This may be an R colour,
a vector of R colours, or
a \code{ggplot2} style "Scale" object e.g. \code{scale_colour_continuous()}.}

\item{selection}{specify the locations in the matrix which will be highlighted.
\describe{
\item{NULL}{(default) Apply highlighting to all elements}
\item{numeric vector}{Numeric vector of indices. e.g. \code{c(1, 2, 3)}}
\item{logical vector}{Either length 1, or a length which matches the
     total number of elements in the matrix/vector e.g. \code{TRUE}}
\item{expression}{An expression which evaluates to a logical vector,
     or vector of indices.
     The matrix/vector itself can be referenced in these expressions
     using the variable \code{.x}
     E.g. \code{abs(.x)> 0.5 & .x != 1}  or \code{row(.x) == 3}}
}}

\item{elem}{Apply the highlighting to the 'fill' (the background) or the 'text'.
Default: 'fill'}

\item{show_legend}{if a scale object is used for colour, and \code{show_legend = TRUE},
then a colourbar legend will be added to the bottom of the output.
Default: FALSE}
}
\description{
\code{hl_mat()} and \code{hl_vec()} are identical functions which both work on
matrices and atomic vectors.
}
