% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_diff.R
\name{hl_str_diff}
\alias{hl_str_diff}
\title{Colour the differences between two strings}
\usage{
hl_str_diff(
  x,
  y,
  bs = "dodgerblue",
  bi = "darkgreen",
  bd = "firebrick",
  fs = "white",
  fi = "white",
  fd = "white",
  ...
)
}
\arguments{
\item{x, y}{each argument is a single string. vectors of strings not currently
supported.}

\item{fs, bs, fi, bi, fd, bd}{the (f)oreground and (b)ackground colouring for
string (s)ubstitutions, (i)insertions and (d)eletions.}

\item{...}{further arguments passed to \code{adist()}}
}
\value{
list of 'emphatic' objects which could be renderd to ANSI or HTML
}
\description{
Highlight the differences between two strings in
terms of substitutions, insertions and deletions.
}
