% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{interp_colour}
\alias{interp_colour}
\title{Interpolate between 2 colour vectors}
\usage{
interp_colour(colour1, colour2, frac)
}
\arguments{
\item{colour1, colour2}{character vectors of R colours}

\item{frac}{fraction linear interpolation between the two}
}
\description{
Interpolate between 2 colour vectors
}
