% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{is_emphatic}
\alias{is_emphatic}
\title{check if data.frame is a emphatic data.frame of matrix}
\usage{
is_emphatic(x)
}
\arguments{
\item{x}{Object to test}
}
\description{
check if data.frame is a emphatic data.frame of matrix
}
