% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{knit_print.emphatic}
\alias{knit_print.emphatic}
\title{Automatically utput emphatic matrices and data.frames to HTML knitted documents.}
\usage{
knit_print.emphatic(x, style = NULL, ...)
}
\arguments{
\item{x}{emphatic object}

\item{style}{html tag styling to apply to the \code{<pre>} wrapper for the
returned HTML}

\item{...}{other arguments passed to \code{as.character.emphatic}}
}
\description{
This function makes use of \code{fansi} package to convert ANSI to HTML.
}
