% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-index-lookup.R
\name{loc_expr_to_ids}
\alias{loc_expr_to_ids}
\title{Convert a column or row expression into a vector of index locations}
\usage{
loc_expr_to_ids(.data, expr, axis)
}
\arguments{
\item{.data}{data.frame}

\item{expr}{an expression wihch will be interpreted as indices}

\item{axis}{'row' or 'column' expression?}
}
\description{
Convert a column or row expression into a vector of index locations
}
