% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-ansi.R
\docType{data}
\name{reset_ansi}
\alias{reset_ansi}
\title{Reset back to terminal defaults}
\format{
An object of class \code{character} of length 1.
}
\usage{
reset_ansi
}
\description{
Reset back to terminal defaults
}
\keyword{datasets}
