% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansi-escape-codes.R
\docType{data}
\name{reset_code}
\alias{reset_code}
\title{Reset back to terminal defaults}
\format{
An object of class \code{character} of length 1.
}
\usage{
reset_code
}
\description{
Reset back to terminal defaults
}
\keyword{datasets}
