% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sets.R
\docType{data}
\name{sea_ice_area}
\alias{sea_ice_area}
\title{Monthly Southern Sea Ice Area over the last 40 years}
\format{
Matrix of sea ice area, monthly from 1978 to 2020.
}
\usage{
sea_ice_area
}
\description{
From the 'National Snow and Ice Data Center' \url{https://nsidc.org/data/g02135}
}
\keyword{datasets}
