% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-html.R
\name{show_html}
\alias{show_html}
\title{Show HTML or SVG content in the rstudio viewer pane}
\usage{
show_html(x, viewer = getOption("viewer", utils::browseURL))
}
\arguments{
\item{x}{svg or html}

\item{viewer}{function which activates viewer}
}
\value{
None
}
\description{
Show HTML or SVG content in the rstudio viewer pane
}
\examples{
\dontrun{
# This example will try and spawn an external viewer for HTML content
# and is thus marked as 'dontrun' so it does not get executed on CRAN
hl_grep(mode, "switch") |>
  as_html() |>
  show_html()
}
}
