% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sets.R
\docType{data}
\name{sydney_rain}
\alias{sydney_rain}
\title{Monthly total rainfall in Centennial Park, Sydney, Australia}
\format{
data.frame with each row representing a year, and each column
        representing a month of that year
}
\usage{
sydney_rain
}
\description{
From the Australian Bureau of Meteorology
}
\keyword{datasets}
