% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-xlsx.R
\name{write_xlsx}
\alias{write_xlsx}
\title{Convert emphatic to Excel workbook}
\usage{
write_xlsx(x, xlsx_filename, colNames = TRUE)
}
\arguments{
\item{x}{emphatic object}

\item{xlsx_filename}{xlsx filename}

\item{colNames}{logical. Default: TRUE}
}
\description{
Convert emphatic to Excel workbook
}
\examples{
mtcars |>
   hl('blue') |>
   write_xlsx(tempfile())
}
