% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{as.character.emphatic}
\alias{as.character.emphatic}
\title{Convert an \code{emphatic} data.frame or matrix to a string.}
\usage{
\method{as.character}{emphatic}(x, ..., mode = "ansi")
}
\arguments{
\item{x}{emphatic data.frame or matrix}

\item{...}{other arguments passed on to \code{format()}}

\item{mode}{Render mode 'ansi' or 'html' determines how the colours will
be represented in text}
}
\description{
The output contains ANSI escape codes to colour the elements in the
data.frame or matrix. This string would then be suitable to pass on to \code{fansi}
for further manipulation e.g. conversion to HTML for displaying in a vignette.
}
