% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-html.R
\name{as_html}
\alias{as_html}
\title{Render an emphatic object to HTML}
\usage{
as_html(x, style = NULL, ..., complete = FALSE)
}
\arguments{
\item{x}{emphatic object}

\item{style}{html tag styling to apply to the \code{<pre>} wrapper for the
returned HTML}

\item{...}{other arguments passed to \code{as.character.emphatic}}

\item{complete}{logical. Default: FALSE.  If TRUE, then add DOCTYPE and
the tags for 'html', 'body' and 'head' to make a complete standalone
html file.}
}
\description{
Render an emphatic object to HTML
}
