% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-latex.R
\name{as_latex}
\alias{as_latex}
\title{Render an emphatic object to Latex}
\usage{
as_latex(x, ..., font_size = NULL)
}
\arguments{
\item{x}{emphatic object}

\item{...}{other arguments passed to \code{as.character.emphatic()}}

\item{font_size}{Integer value indicating font size measured in points.
Default: NULL.}
}
\value{
single character string containing a latex representation
}
\description{
Render an emphatic object to Latex
}
\examples{
hl_diff("hello", "there") |>
  as_latex() |>
  cat()
}
