% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-svg.R
\name{as_svg}
\alias{as_svg}
\title{Wrap a single emphatic object into an SVG for display}
\usage{
as_svg(x, ..., width = 1200, height = 900, browsable = FALSE)
}
\arguments{
\item{x}{emphatic object}

\item{...}{other arguments passed to \code{as.character.emphatic()}}

\item{width, height}{viewBox dimensions for SVG}

\item{browsable}{Should the SVG be rendered to the RStudio Viewer pane when
when printed (instead of console output)? Default: FALSE}
}
\value{
character string containing an SVG snippet.

Character string containing SVG representation
}
\description{
This is mainly useful within a github \code{README.md} since github will
not rendered html-styled text in colour, but \emph{will} render it correctly
if it is within a \code{<svg>} tags.
}
\details{
This is just a the results of \code{as_html()} wrapped in \code{<svg>} tags
}
\examples{
hl_diff('hello', 'there') |> as_svg()
}
