% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-html.R
\name{as_svg_anim}
\alias{as_svg_anim}
\title{Wrap an emphatic object into an SVG for display in a github README.md}
\usage{
as_svg_anim(..., width = 1200, height = 900, duration = 1)
}
\arguments{
\item{...}{multiple emphatic objects}

\item{width, height}{viewBox dimensions for SVG}

\item{duration}{frame duration in seconds. May be a single value used for
all frames, or a vector of values (one duration value for each frame).
Can be fractions of a second.}
}
\value{
character string containing an SVG snippet.
}
\description{
Idea borrowed from pointblank
}
