% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-html.R
\name{as_svg_anim}
\alias{as_svg_anim}
\title{Wrap multiple emphatic object into an SVG animation}
\usage{
as_svg_anim(
  x,
  width = 1200,
  height = 900,
  duration = 1,
  playback = c("infinite", "click"),
  svg_id = NULL
)
}
\arguments{
\item{x}{list of emphatic objects}

\item{width, height}{viewBox dimensions for SVG}

\item{duration}{frame duration in seconds. May be a single value used for
all frames, or a vector of values (one duration value for each frame).
Can be fractions of a second.}

\item{playback}{'click', 'infinite'}
}
\value{
character string containing an SVG snippet.
}
\description{
Idea borrowed from pointblank
}
