% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-html.R
\name{as_svg_group}
\alias{as_svg_group}
\title{Wrap an emphatic object into an SVG for display in a github README.md}
\usage{
as_svg_group(x, width = 1200, height = 900, ...)
}
\arguments{
\item{x}{emphatic object}

\item{width, height}{viewBox dimensions for SVG}

\item{...}{other arguments passed to \code{as.character.emphatic}}
}
\value{
character string containing an SVG snippet.
}
\description{
Idea borrowed from pointblank
}
