% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{is_emphatic}
\alias{is_emphatic}
\title{Check if data.frame, matrix or atomic vector is a valid emphatic version}
\usage{
is_emphatic(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical value
}
\description{
Check if data.frame, matrix or atomic vector is a valid emphatic version
}
\examples{
mtcars |>
  hl('red') |>
  is_emphatic()
}
