% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{knit_print.emphatic}
\alias{knit_print.emphatic}
\title{Automatically output emphatic objects to HTML knitted documents.}
\usage{
knit_print.emphatic(x, style = list(), ...)
}
\arguments{
\item{x}{emphatic object}

\item{style}{html tag styling to apply to the \code{<pre>} wrapper for the
returned HTML}

\item{...}{other arguments passed to \code{as.character.emphatic()}}
}
\value{
a character vector suitable for output during an rmarkdown render
}
\description{
Automatically output emphatic objects to HTML knitted documents.
}
\examples{
mtcars |>
  hl('red') |>
  knit_print.emphatic()
}
