% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{print.emphatic}
\alias{print.emphatic}
\title{Print an \code{emphatic} data.frame, matrix or atomic vector}
\usage{
\method{print}{emphatic}(x, ...)
}
\arguments{
\item{x}{\code{emphatic} data.frame, matrix or atomic vector}

\item{...}{other arguments passed on to \code{format()}}
}
\value{
None.
}
\description{
Print an \code{emphatic} data.frame, matrix or atomic vector
}
\examples{
mtcars |>
  head() |>
  hl('red') |>
  print()
}
