% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-xlsx.R
\name{write_xlsx}
\alias{write_xlsx}
\title{Write an emphatic data.frame to an Excel workbook}
\usage{
write_xlsx(x, xlsx_filename, colNames = TRUE, opts = hl_opts())
}
\arguments{
\item{x}{emphatic data.frame object}

\item{xlsx_filename}{xlsx filename}

\item{colNames}{Display column names? logical. Default: TRUE}

\item{opts}{rendering options}
}
\value{
None
}
\description{
Requires \code{openxlsx} package
}
\examples{
mtcars |>
   hl('blue') |>
   write_xlsx(tempfile())
}
